fix(http): Refactor config.Compression and use it in http
